Un-extract the ExpressionBuilder class into SupergrammarSpanner.
